# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bp-carnet",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/password-reset": {
      "post": {
        "summary": "再設定を受け付け、メール確認済みのアカウントがあれば再設定のリンクを、確認前のアカウント\nなら確認リンクを送る。",
        "description": "応答はアカウントの有無に関わらず同じにする。アカウントが無いアドレスにはメールを送らない\n(無関係なアドレスへの送りつけに使わせないため)。",
        "operationId": "request_password_reset",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "受け付けた (アカウントが無いメールアドレスでも同じ応答)"
          },
          "400": {
            "description": "リクエストボディが不正 (JSON として解釈できない等)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Content-Type が application/json でない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "メールアドレスとして解釈できない、または必須フィールドが無い",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "同じ接続元からの要求が上限を超えた",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/password-reset/check": {
      "post": {
        "summary": "再設定のリンクがまだ使えるかを、使用済みにせずに確かめる (リンクを開いたときに使う)。",
        "operationId": "check_password_reset_link",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "リンクは使える"
          },
          "400": {
            "description": "リンクが無効・期限切れ・使用済み、またはリクエストボディが不正",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Content-Type が application/json でない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "必須フィールドが無い",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/password-reset/complete": {
      "post": {
        "summary": "リンクの先で新しいパスワードを設定する。そのアカウントのセッションはすべて断つ。",
        "description": "ADR: ログインはさせず、ログイン画面から入り直してもらう。凍結中の拒否や削除予約の取り消しと\nいったログイン時の判定を、ログインの経路だけに保つため。",
        "operationId": "complete_password_reset",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompletePasswordResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "再設定した (ログインはしない)"
          },
          "400": {
            "description": "リンクが無効・期限切れ・使用済み、またはリクエストボディが不正",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Content-Type が application/json でない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "パスワードがポリシーを満たさない、または必須フィールドが無い",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "同じ接続元、または全体からの要求が上限を超えた",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckLinkRequest": {
        "type": "object",
        "description": "リンクの先を開いたときに、リンクがまだ使えるかを尋ねる要求。",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string",
            "description": "メールのリンクに載っていたトークン。"
          }
        }
      },
      "CompletePasswordResetRequest": {
        "type": "object",
        "required": [
          "token",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          },
          "token": {
            "type": "string",
            "description": "再設定のメールのリンクに載っていたトークン。"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "rows": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/RowError"
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "レスポンスボディの共通 envelope。`code` が機械可読な契約 (frontend はこちらで表示文言を引く)、\n`message` はローカライズしないデバッグ用の説明文。`rows` は `BatchValidation` のときだけ埋まる。",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "PasswordResetRequest": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string",
            "example": "kyoko@example.com"
          }
        }
      },
...
      "RowError": {
        "type": "object",
        "description": "バッチ処理で 1 行の検証が失敗したことを表す。`index` はリクエスト配列内の 0 始まりの位置。",
        "required": [
          "index",
          "message"
        ],
        "properties": {
          "index": {
            "type": "integer",
            "minimum": 0
          },
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  }
}